refactor(domain): domain 패키지로 묶음
