+++ src/main/java/janggi/JanggiGame.java
@@ -1,9 +1,9 @@
 package janggi;
 
-import janggi.board.JanggiBoard;
-import janggi.board.Position;
-import janggi.piece.Piece;
-import janggi.piece.Side;
+import janggi.domain.board.JanggiBoard;
+import janggi.domain.board.Position;
+import janggi.domain.piece.Piece;
+import janggi.domain.piece.Side;
 import janggi.view.InputView;
 import janggi.view.OutputView;
