+++ src/main/java/janggi/domain/board/BoardInitializer.java
@@ -1,30 +1,30 @@
-package janggi.board;
+package janggi.domain.board;
 
-import static janggi.board.InitialPositions.CHO_CANNON_POSITIONS;
-import static janggi.board.InitialPositions.CHO_CHARIOT_POSITIONS;
-import static janggi.board.InitialPositions.CHO_ELEPHANT_POSITIONS;
-import static janggi.board.InitialPositions.CHO_GUARD_POSITIONS;
-import static janggi.board.InitialPositions.CHO_HORSE_POSITIONS;
-import static janggi.board.InitialPositions.CHO_KING_POSITIONS;
-import static janggi.board.InitialPositions.CHO_SOLDIER_POSITIONS;
-import static janggi.board.InitialPositions.HAN_CANNON_POSITIONS;
-import static janggi.board.InitialPositions.HAN_CHARIOT_POSITIONS;
-import static janggi.board.InitialPositions.HAN_ELEPHANT_POSITIONS;
-import static janggi.board.InitialPositions.HAN_GUARD_POSITIONS;
-import static janggi.board.InitialPositions.HAN_HORSE_POSITIONS;
-import static janggi.board.InitialPositions.HAN_KING_POSITIONS;
-import static janggi.board.InitialPositions.HAN_SOLDIER_POSITIONS;
+import static janggi.domain.board.InitialPositions.CHO_CANNON_POSITIONS;
+import static janggi.domain.board.InitialPositions.CHO_CHARIOT_POSITIONS;
+import static janggi.domain.board.InitialPositions.CHO_ELEPHANT_POSITIONS;
+import static janggi.domain.board.InitialPositions.CHO_GUARD_POSITIONS;
+import static janggi.domain.board.InitialPositions.CHO_HORSE_POSITIONS;
+import static janggi.domain.board.InitialPositions.CHO_KING_POSITIONS;
+import static janggi.domain.board.InitialPositions.CHO_SOLDIER_POSITIONS;
+import static janggi.domain.board.InitialPositions.HAN_CANNON_POSITIONS;
+import static janggi.domain.board.InitialPositions.HAN_CHARIOT_POSITIONS;
+import static janggi.domain.board.InitialPositions.HAN_ELEPHANT_POSITIONS;
+import static janggi.domain.board.InitialPositions.HAN_GUARD_POSITIONS;
+import static janggi.domain.board.InitialPositions.HAN_HORSE_POSITIONS;
+import static janggi.domain.board.InitialPositions.HAN_KING_POSITIONS;
+import static janggi.domain.board.InitialPositions.HAN_SOLDIER_POSITIONS;
 
-import janggi.piece.Empty;
-import janggi.piece.Piece;
-import janggi.piece.Side;
-import janggi.piece.limit.Elephant;
-import janggi.piece.limit.Guard;
-import janggi.piece.limit.Horse;
-import janggi.piece.limit.King;
-import janggi.piece.limit.Soldier;
-import janggi.piece.unlimit.Cannon;
-import janggi.piece.unlimit.Chariot;
+import janggi.domain.piece.Empty;
+import janggi.domain.piece.Piece;
+import janggi.domain.piece.Side;
+import janggi.domain.piece.limit.Elephant;
+import janggi.domain.piece.limit.Guard;
+import janggi.domain.piece.limit.Horse;
+import janggi.domain.piece.limit.King;
+import janggi.domain.piece.limit.Soldier;
+import janggi.domain.piece.unlimit.Cannon;
+import janggi.domain.piece.unlimit.Chariot;
 import java.util.HashMap;
 import java.util.Map;
