+++ src/main/java/janggi/domain/board/InitialPositions.java
@@ -1,4 +1,4 @@
-package janggi.board;
+package janggi.domain.board;
 
 import java.util.Collections;
 import java.util.List;
