+++ src/main/java/janggi/domain/board/PalacePoint.java
@@ -1,4 +1,4 @@
-package janggi.board;
+package janggi.domain.board;
 
 import java.util.ArrayList;
 import java.util.Arrays;
