+++ src/main/java/janggi/domain/piece/Empty.java
@@ -1,7 +1,7 @@
-package janggi.piece;
+package janggi.domain.piece;
 
-import janggi.board.Position;
-import janggi.move.Route;
+import janggi.domain.board.Position;
+
 import java.util.List;
 import java.util.Map;
