+++ src/main/java/janggi/domain/piece/Piece.java
@@ -1,6 +1,6 @@
-package janggi.piece;
+package janggi.domain.piece;
 
-import janggi.board.Position;
+import janggi.domain.board.Position;
 import java.util.List;
 import java.util.Map;
