+++ src/main/java/janggi/domain/piece/PieceType.java
@@ -1,4 +1,4 @@
-package janggi.piece;
+package janggi.domain.piece;
 
 public enum PieceType {
     KING("G", 0),
