+++ src/main/java/janggi/domain/piece/limit/Elephant.java
@@ -1,17 +1,17 @@
-package janggi.piece.limit;
+package janggi.domain.piece.limit;
 
-import janggi.board.Position;
-import janggi.move.Direction;
-import janggi.move.Route;
-import janggi.piece.PieceType;
-import janggi.piece.Side;
+import janggi.domain.board.Position;
+import janggi.domain.move.Direction;
+import janggi.domain.move.Route;
+import janggi.domain.piece.PieceType;
+import janggi.domain.piece.Side;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class Horse extends LimitMovable {
+public class Elephant extends LimitMovable {
 
-    public Horse(final Side side) {
+    public Elephant(final Side side) {
         super(side);
     }
 
@@ -20,12 +20,12 @@ public class Horse extends LimitMovable {
         List<Route> movableRoutes = new ArrayList<>();
         movableRoutes.addAll(createRoute(position, Direction.UP, Direction.LEFT_UP));
         movableRoutes.addAll(createRoute(position, Direction.UP, Direction.RIGHT_UP));
+        movableRoutes.addAll(createRoute(position, Direction.DOWN, Direction.LEFT_DOWN));
+        movableRoutes.addAll(createRoute(position, Direction.DOWN, Direction.RIGHT_DOWN));
         movableRoutes.addAll(createRoute(position, Direction.LEFT, Direction.LEFT_UP));
         movableRoutes.addAll(createRoute(position, Direction.LEFT, Direction.LEFT_DOWN));
         movableRoutes.addAll(createRoute(position, Direction.RIGHT, Direction.RIGHT_UP));
         movableRoutes.addAll(createRoute(position, Direction.RIGHT, Direction.RIGHT_DOWN));
-        movableRoutes.addAll(createRoute(position, Direction.DOWN, Direction.LEFT_DOWN));
-        movableRoutes.addAll(createRoute(position, Direction.DOWN, Direction.RIGHT_DOWN));
 
         movableRoutes.removeIf(route -> route.getPositions().isEmpty());
         return movableRoutes;
@@ -35,9 +35,10 @@ public class Horse extends LimitMovable {
                               final Direction diagonalNormalDirection) {
         Position movedPosition = originalPosition.move(normalDirection);
         Position diagonalMovedPosition = movedPosition.move(diagonalNormalDirection);
+        Position diagonalMovedPosition2 = diagonalMovedPosition.move(diagonalNormalDirection);
 
-        if (movedPosition.isInBoardRange() && diagonalMovedPosition.isInBoardRange()) {
-            return List.of(new Route(movedPosition, diagonalMovedPosition));
+        if(movedPosition.isInBoardRange() && diagonalMovedPosition.isInBoardRange() && diagonalMovedPosition2.isInBoardRange()){
+            return List.of(new Route(movedPosition, diagonalMovedPosition, diagonalMovedPosition2));
         }
 
         return List.of();
@@ -45,6 +46,6 @@ public class Horse extends LimitMovable {
 
     @Override
     public PieceType getType() {
-        return PieceType.HORSE;
+        return PieceType.ELEPHANT;
     }
 }
