+++ src/main/java/janggi/domain/piece/limit/Horse.java
@@ -1,17 +1,17 @@
-package janggi.piece.limit;
+package janggi.domain.piece.limit;
 
-import janggi.board.Position;
-import janggi.move.Direction;
-import janggi.move.Route;
-import janggi.piece.PieceType;
-import janggi.piece.Side;
+import janggi.domain.board.Position;
+import janggi.domain.move.Direction;
+import janggi.domain.move.Route;
+import janggi.domain.piece.PieceType;
+import janggi.domain.piece.Side;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class Elephant extends LimitMovable {
+public class Horse extends LimitMovable {
 
-    public Elephant(final Side side) {
+    public Horse(final Side side) {
         super(side);
     }
 
@@ -20,12 +20,12 @@ public class Elephant extends LimitMovable {
         List<Route> movableRoutes = new ArrayList<>();
         movableRoutes.addAll(createRoute(position, Direction.UP, Direction.LEFT_UP));
         movableRoutes.addAll(createRoute(position, Direction.UP, Direction.RIGHT_UP));
-        movableRoutes.addAll(createRoute(position, Direction.DOWN, Direction.LEFT_DOWN));
-        movableRoutes.addAll(createRoute(position, Direction.DOWN, Direction.RIGHT_DOWN));
         movableRoutes.addAll(createRoute(position, Direction.LEFT, Direction.LEFT_UP));
         movableRoutes.addAll(createRoute(position, Direction.LEFT, Direction.LEFT_DOWN));
         movableRoutes.addAll(createRoute(position, Direction.RIGHT, Direction.RIGHT_UP));
         movableRoutes.addAll(createRoute(position, Direction.RIGHT, Direction.RIGHT_DOWN));
+        movableRoutes.addAll(createRoute(position, Direction.DOWN, Direction.LEFT_DOWN));
+        movableRoutes.addAll(createRoute(position, Direction.DOWN, Direction.RIGHT_DOWN));
 
         movableRoutes.removeIf(route -> route.getPositions().isEmpty());
         return movableRoutes;
@@ -35,10 +35,9 @@ public class Elephant extends LimitMovable {
                               final Direction diagonalNormalDirection) {
         Position movedPosition = originalPosition.move(normalDirection);
         Position diagonalMovedPosition = movedPosition.move(diagonalNormalDirection);
-        Position diagonalMovedPosition2 = diagonalMovedPosition.move(diagonalNormalDirection);
 
-        if(movedPosition.isInBoardRange() && diagonalMovedPosition.isInBoardRange() && diagonalMovedPosition2.isInBoardRange()){
-            return List.of(new Route(movedPosition, diagonalMovedPosition, diagonalMovedPosition2));
+        if (movedPosition.isInBoardRange() && diagonalMovedPosition.isInBoardRange()) {
+            return List.of(new Route(movedPosition, diagonalMovedPosition));
         }
 
         return List.of();
@@ -46,6 +45,6 @@ public class Elephant extends LimitMovable {
 
     @Override
     public PieceType getType() {
-        return PieceType.ELEPHANT;
+        return PieceType.HORSE;
     }
 }
