+++ src/test/java/janggi/piece/limit/SoldierTest.java
@@ -1,8 +1,9 @@
 package janggi.piece.limit;
 
-import janggi.board.Position;
-import janggi.move.Route;
-import janggi.piece.Side;
+import janggi.domain.board.Position;
+import janggi.domain.move.Route;
+import janggi.domain.piece.Side;
+import janggi.domain.piece.limit.Soldier;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
